+++ src/test/java/ar/edu/unq/desapp/GrupoA022022/backenddesappapi/model/OperationModelTests.java
@@ -1,9 +1,6 @@
-package ar.edu.unq.desapp.GrupoA022022.backenddesappapi;
+package ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model;
 
-import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Cryptocurrency;
-import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Intention;
-import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Operation;
-import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.User;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.DataSet;
 import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.persistence.ICryptocurrencyRepo;
 import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.persistence.IIntentionRepo;
 import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.persistence.IOperationRepo;
@@ -12,13 +9,12 @@ import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.utils.IntentionType;
 import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.utils.OperationState;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest
-@AutoConfigureTestDatabase
-class OperationTests {
+class OperationModelTests {
 
     DataSet dataSet = new DataSet();
 
@@ -64,20 +60,5 @@ class OperationTests {
         Operation operation = new Operation(dataSet.getIntentionBuy(), dataSet.getUserTest());
 
         assertEquals(0, operation.getUserReputation());
-    }
-
-    //**************** SERVICE - REPOSITORY ****************
-
-    @Test
-    void recoversPersistanceANewOperation() {
-        User someuserDB = userRepo.save(dataSet.getUserTest());
-        User someuser2DB = userRepo.save(dataSet.getUserTest2());
-        Cryptocurrency somecryptocurrencyDB = cryptocurrencyRepo.save(dataSet.getCryptocurrency4());
-        Intention intentionDB = intentionRepo.save(new Intention(dataSet.getSomeType(), somecryptocurrencyDB,
-                dataSet.getSomePrice(), dataSet.getSomeUnit(), someuserDB));
-        Operation saved = operationRepo.save(new Operation(intentionDB,someuser2DB));
-
-        int idSaved = saved.getId();
-        assertEquals(operationRepo.findById(idSaved).get().getId(), idSaved);
     }
 }
