+++ src/test/java/ar/edu/unq/desapp/GrupoA022022/backenddesappapi/persistence/OperationPersistenceTests.java
@@ -1,0 +1,48 @@
+package ar.edu.unq.desapp.GrupoA022022.backenddesappapi.persistence;
+
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.DataSet;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Cryptocurrency;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Intention;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.Operation;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.model.User;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.utils.IntentionType;
+import ar.edu.unq.desapp.GrupoA022022.backenddesappapi.utils.OperationState;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@SpringBootTest
+@AutoConfigureTestDatabase
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class OperationPersistenceTests {
+
+    DataSet dataSet = new DataSet();
+
+    @Autowired
+    IIntentionRepo intentionRepo;
+    @Autowired
+    IOperationRepo operationRepo;
+    @Autowired
+    IUserRepo userRepo;
+    @Autowired
+    ICryptocurrencyRepo cryptocurrencyRepo;
+
+    //**************** SERVICE - REPOSITORY ****************
+
+    @Test
+    void recoversPersistanceANewOperation() {
+        User someuserDB = userRepo.save(dataSet.getUserTest());
+        User someuser2DB = userRepo.save(dataSet.getUserTest2());
+        Cryptocurrency somecryptocurrencyDB = cryptocurrencyRepo.save(dataSet.getCryptocurrency4());
+        Intention intentionDB = intentionRepo.save(new Intention(dataSet.getSomeType(), somecryptocurrencyDB,
+                dataSet.getSomePrice(), dataSet.getSomeUnit(), someuserDB));
+        Operation saved = operationRepo.save(new Operation(intentionDB,someuser2DB));
+
+        int idSaved = saved.getId();
+        assertEquals(operationRepo.findById(idSaved).get().getId(), idSaved);
+    }
+}
